+++ Projeto/src/main/java/br/com/PortalRh/Projeto/repository/PositionRepository.java
@@ -1,6 +1,5 @@
 package br.com.PortalRh.Projeto.repository;
 
-import br.com.PortalRh.Projeto.model.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
