+++ Projeto/src/main/java/br/com/PortalRh/Projeto/resource/CertificationController.java
@@ -1,4 +1,4 @@
-package br.com.PortalRh.Projeto.resourse;
+package br.com.PortalRh.Projeto.resource;
 
 import br.com.PortalRh.Projeto.entities.Certification;
 import br.com.PortalRh.Projeto.service.CertificationService;
