+++ Projeto/src/main/java/br/com/PortalRh/Projeto/resource/ChildrenController.java
@@ -1,4 +1,4 @@
-package br.com.PortalRh.Projeto.resourse;
+package br.com.PortalRh.Projeto.resource;
 
 import br.com.PortalRh.Projeto.entities.Children;
 import br.com.PortalRh.Projeto.service.ChildrenService;
