+++ Projeto/src/main/java/br/com/PortalRh/Projeto/resource/UserController.java
@@ -1,4 +1,4 @@
-package br.com.PortalRh.Projeto.resourse;
+package br.com.PortalRh.Projeto.resource;
 
 import br.com.PortalRh.Projeto.model.Filiacao;
 import br.com.PortalRh.Projeto.model.User;
